+++ CHANGELOG.md
@@ -12,6 +12,7 @@ and this project adheres to [Semantic Versioning](http://semver.org/).
 * add a formatter function for printer.go
 * add `priority:` as an input when adding a task
 * add CreatedAt and UpdatedAt field in Task struct
+* add the possibility to list task by projects
 
 ### Changed
 * change priority from bool to int to handle priority levels